LayoutTests/media/modern-media-controls/tracks-support/no-tracks-expected.txt

Testing the TracksSupport behavior with no tracks.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS shadowRoot.querySelector('button.tracks').getBoundingClientRect().width is 0

PASS successfullyParsed is true

TEST COMPLETE
